refactor(data): モックデータを保存する処理を共通化
