+++ internal/infrastructure/rdb/plan_candidate_test.go
@@ -2,10 +2,7 @@ package rdb
 
 import (
 	"context"
-	"database/sql"
-	"fmt"
 	"github.com/google/uuid"
-	"github.com/volatiletech/null/v8"
 	"github.com/volatiletech/sqlboiler/v4/boil"
 	"github.com/volatiletech/sqlboiler/v4/queries/qm"
 	"poroto.app/poroto/planner/internal/domain/array"
@@ -172,7 +169,7 @@ func TestPlanCandidateRepository_Find(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -263,7 +260,7 @@ func TestPlanCandidateRepository_Find_ShouldReturnNil(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -334,7 +331,7 @@ func TestPlanCandidateRepository_FindPlan(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -416,7 +413,7 @@ func TestPlanCandidateRepository_FindExpiredBefore(t *testing.T) {
 
 			// 事前にPlanCandidateSetを作成しておく
 			for _, planCandidateSet := range c.savedPlanCandidateSets {
-				if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, planCandidateSet); err != nil {
+				if err := savePlanCandidate(testContext, testDB, planCandidateSet); err != nil {
 					t.Fatalf("failed to save plan candidate: %v", err)
 				}
 			}
@@ -468,7 +465,7 @@ func TestPlanCandidateRepository_AddSearchedPlacesForPlanCandidate(t *testing.T)
 			})
 
 			// 事前にPlanCandidateSetを作成しておく]
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, models.PlanCandidate{Id: c.planCandidateId, ExpiresAt: time.Now().Add(time.Hour)}); err != nil {
+			if err := savePlanCandidate(testContext, testDB, models.PlanCandidate{Id: c.planCandidateId, ExpiresAt: time.Now().Add(time.Hour)}); err != nil {
 				t.Fatalf("failed to create plan candidate: %v", err)
 			}
 
@@ -549,7 +546,7 @@ func TestPlanCandidateRepository_AddPlan(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, models.PlanCandidate{Id: c.planCandidateId, ExpiresAt: time.Now().Add(time.Hour)}); err != nil {
+			if err := savePlanCandidate(testContext, testDB, models.PlanCandidate{Id: c.planCandidateId, ExpiresAt: time.Now().Add(time.Hour)}); err != nil {
 				t.Fatalf("failed to create plan candidate: %v", err)
 			}
 
@@ -747,7 +744,7 @@ func TestPlanCandidateRepository_RemovePlaceFromPlan(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -822,7 +819,7 @@ func TestPlanCandidateRepository_UpdatePlacesOrder(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -900,7 +897,7 @@ func TestPlanCandidateRepository_UpdatePlacesOrder_ShouldReturnError(t *testing.
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -975,7 +972,7 @@ func TestPlanCandidateRepository_UpdatePlanCandidateMetaData(t *testing.T) {
 			})
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -1093,7 +1090,7 @@ func TestPlanCandidateRepository_ReplacePlace(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -1174,7 +1171,7 @@ func TestPlanCandidateRepository_ReplacePlace_ShouldReturnError(t *testing.T) {
 			}
 
 			// 事前にPlanCandidateSetを作成しておく
-			if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, c.savedPlanCandidateSet); err != nil {
+			if err := savePlanCandidate(testContext, testDB, c.savedPlanCandidateSet); err != nil {
 				t.Fatalf("failed to save plan candidate: %v", err)
 			}
 
@@ -1270,7 +1267,7 @@ func TestPlanCandidateRepository_DeleteAll(t *testing.T) {
 
 			// 事前にPlanCandidateSetを作成しておく
 			for _, planCandidateSet := range c.savedPlanCandidateSets {
-				if err := savePlanCandidate(testContext, testDB, *planCandidateRepository, planCandidateSet); err != nil {
+				if err := savePlanCandidate(testContext, testDB, planCandidateSet); err != nil {
 					t.Fatalf("failed to save plan candidate: %v", err)
 				}
 			}
@@ -1363,60 +1360,6 @@ func TestPlanCandidateRepository_DeleteAll(t *testing.T) {
 			}
 		})
 	}
-}
-
-func savePlaces(ctx context.Context, db *sql.DB, places []models.Place) error {
-	places = array.DistinctBy(places, func(place models.Place) string { return place.Id })
-	for _, place := range places {
-		placeEntity := entities.Place{ID: place.Id}
-		if err := placeEntity.Insert(ctx, db, boil.Infer()); err != nil {
-			return fmt.Errorf("failed to insert place: %v", err)
-		}
-
-		if place.Google.PlaceId == "" {
-			continue
-		}
-
-		googlePlaceEntity := entities.GooglePlace{PlaceID: place.Google.PlaceId}
-		if err := placeEntity.AddGooglePlaces(ctx, db, true, &googlePlaceEntity); err != nil {
-			return fmt.Errorf("failed to insert google place: %v", err)
-		}
-	}
-
-	return nil
-}
-
-func savePlanCandidate(ctx context.Context, db *sql.DB, planCandidateRepository PlanCandidateRepository, planCandidateSet models.PlanCandidate) error {
-	// PlanCandidateSetを作成
-	planCandidateSetEntity := entities.PlanCandidateSet{
-		ID:        planCandidateSet.Id,
-		ExpiresAt: planCandidateSet.ExpiresAt,
-	}
-	if err := planCandidateSetEntity.Insert(ctx, db, boil.Infer()); err != nil {
-		return fmt.Errorf("failed to insert plan candidate set: %v", err)
-	}
-
-	// PlanCandidateSetMetaDataを作成
-	if !planCandidateSet.MetaData.IsZero() {
-		planCandidateSetMetaDataEntity := entities.PlanCandidateSetMetaDatum{
-			ID:                           uuid.New().String(),
-			PlanCandidateSetID:           planCandidateSet.Id,
-			IsCreatedFromCurrentLocation: planCandidateSet.MetaData.CreatedBasedOnCurrentLocation,
-			LatitudeStart:                planCandidateSet.MetaData.LocationStart.Latitude,
-			LongitudeStart:               planCandidateSet.MetaData.LocationStart.Longitude,
-			PlanDurationMinutes:          null.IntFromPtr(planCandidateSet.MetaData.FreeTime),
-		}
-		if err := planCandidateSetMetaDataEntity.Insert(ctx, db, boil.Infer()); err != nil {
-			return fmt.Errorf("failed to insert plan candidate set meta data: %v", err)
-		}
-	}
-
-	// PlanCandidateを作成
-	if err := planCandidateRepository.AddPlan(ctx, planCandidateSet.Id, planCandidateSet.Plans...); err != nil {
-		return fmt.Errorf("failed to add plan: %v", err)
-	}
-
-	return nil
 }
 
 func toPointer[T any](value T) *T {
